Fix warnings in Rotator ctors

+++ EngineMath/Rotator.cpp
@@ -13,26 +13,26 @@ namespace EngineMath {
 	Rotator::Rotator(float pitch, float yaw, float roll) {
 		//Bound pitch -360 -> 360 (exclusive)
 		if (pitch > 359) {
-			pitch = std::fmod(pitch, 360);
+			pitch = (float)std::fmod(pitch, 360);
 		}
 		if (pitch < -359) {
-			pitch = std::fmod(pitch, -360);
+			pitch = (float)std::fmod(pitch, -360);
 		}
 
 		//Bound yaw -360 -> 360 (exclusive)
 		if (yaw > 359) {
-			yaw = std::fmod(yaw, 360);
+			yaw = (float)std::fmod(yaw, 360);
 		}
 		if (pitch < -359) {
-			yaw = std::fmod(yaw, -360);
+			yaw = (float)std::fmod(yaw, -360);
 		}
 
 		//Bound roll -360 -> 360 (exclusive)
 		if (roll > 359) {
-			roll = std::fmod(roll, 360);
+			roll = (float)std::fmod(roll, 360);
 		}
 		if (roll < -359) {
-			roll = std::fmod(roll, -360);
+			roll = (float)std::fmod(roll, -360);
 		}
 
 		Pitch = pitch;
